requirements.txt
Django==5.1.6
Pillow==11.2.1
PyMySQL==1.1.1
